extracted superpose routines
